# api/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Elmanhg.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/subjects/{subjectId}/units/{unitId}": {
      "put": {
        "tags": [
          "Units"
        ],
        "operationId": "UpdateUnit",
        "parameters": [
          {
            "name": "subjectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "unitId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitNameRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitNameRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UnitNameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Units"
        ],
        "operationId": "DeleteUnit",
        "parameters": [
          {
            "name": "subjectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "unitId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/subjects/{subjectId}": {
      "get": {
        "tags": [
          "Subjects"
        ],
        "operationId": "GetSubject",
        "parameters": [
          {
            "name": "subjectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDetailResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDetailResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDetailResult"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Subjects"
        ],
        "operationId": "UpdateSubject",
        "parameters": [
          {
            "name": "subjectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectNameRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectNameRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectNameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Subjects"
        ],
        "operationId": "DeleteSubject",
        "parameters": [
          {
            "name": "subjectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/lessons/{lessonId}": {
      "get": {
        "tags": [
          "Lessons"
        ],
        "operationId": "GetLesson",
        "parameters": [
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDetailResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDetailResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDetailResult"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Lessons"
        ],
        "operationId": "UpdateLesson",
        "parameters": [
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLessonRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLessonRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLessonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Lessons"
        ],
        "operationId": "DeleteLesson",
        "parameters": [
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/browse/subjects/{subjectId}": {
      "get": {
        "tags": [
          "Browse"
        ],
        "operationId": "GetStudentSubject",
        "parameters": [
          {
            "name": "subjectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentSubjectResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentSubjectResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentSubjectResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/browse/units/{unitId}": {
      "get": {
        "tags": [
          "Browse"
        ],
        "operationId": "GetStudentUnit",
        "parameters": [
          {
            "name": "unitId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentUnitResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentUnitResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentUnitResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/browse/lessons/{lessonId}": {
      "get": {
        "tags": [
          "Browse"
        ],
        "operationId": "GetStudentLesson",
        "parameters": [
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentLessonResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentLessonResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentLessonResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/browse/lessons/{lessonId}/openings": {
      "post": {
        "tags": [
          "Browse"
        ],
        "operationId": "RecordLessonOpening",
        "parameters": [
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LessonDetailResult": {
        "required": [
          "id",
          "unitId",
          "name",
          "order",
          "state",
          "explanation",
          "summary",
          "videoUrl",
          "objectives"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "unitId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "order": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "state": {
            "type": "string"
          },
          "explanation": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "videoUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "objectives": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LessonObjectiveResult"
            }
          }
        }
      },
      "LessonLinkResult": {
        "required": [
          "id",
          "name",
          "unitId",
          "unitName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "unitId": {
            "type": "string",
            "format": "uuid"
          },
          "unitName": {
            "type": "string"
          }
        }
      },
...
      "LessonObjectiveRequest": {
        "required": [
          "id",
          "text"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "LessonObjectiveResult": {
        "required": [
          "id",
          "text",
          "order"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string"
          },
          "order": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "StudentLessonResult": {
        "required": [
          "id",
          "name",
          "unitId",
          "unitName",
          "subjectId",
          "subjectName",
          "explanation",
          "summary",
          "videoUrl",
          "objectives",
          "servableCount",
          "masteredCount",
          "seenCount",
          "masteryPercent",
          "previousLesson",
          "nextLesson"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "unitId": {
            "type": "string",
            "format": "uuid"
          },
          "unitName": {
            "type": "string"
          },
          "subjectId": {
            "type": "string",
            "format": "uuid"
          },
          "subjectName": {
            "type": "string"
          },
          "explanation": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "videoUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "objectives": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LessonObjectiveResult"
            }
          },
          "servableCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "masteredCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "seenCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "masteryPercent": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "previousLesson": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LessonLinkResult"
              }
            ]
          },
          "nextLesson": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LessonLinkResult"
              }
            ]
          }
        }
      },
      "StudentLessonSummaryResult": {
        "required": [
          "id",
          "name",
          "servableCount",
          "masteredCount",
          "seenCount",
          "masteryPercent"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "servableCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "masteredCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "seenCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "masteryPercent": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "StudentSubjectResult": {
        "required": [
          "id",
          "name",
          "servableCount",
          "masteredCount",
          "seenCount",
          "masteryPercent",
          "units"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "servableCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "masteredCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "seenCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "masteryPercent": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "units": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentUnitSummaryResult"
            }
          }
        }
      },
      "StudentUnitResult": {
        "required": [
          "id",
          "name",
          "subjectId",
          "subjectName",
          "servableCount",
          "masteredCount",
          "seenCount",
          "masteryPercent",
          "bestExamScorePercent",
          "lessons"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "subjectId": {
            "type": "string",
            "format": "uuid"
          },
          "subjectName": {
            "type": "string"
          },
          "servableCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "masteredCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "seenCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "masteryPercent": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "bestExamScorePercent": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "lessons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentLessonSummaryResult"
            }
          }
        }
      },
      "StudentUnitSummaryResult": {
        "required": [
          "id",
          "name",
          "lessonCount",
          "servableCount",
          "masteredCount",
          "seenCount",
          "masteryPercent",
          "bestExamScorePercent"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "lessonCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "servableCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "masteredCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "seenCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "masteryPercent": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "bestExamScorePercent": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "SubjectDetailResult": {
        "required": [
          "id",
          "name",
          "order",
          "units"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "order": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "units": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnitResult"
            }
          }
        }
      },
...
      "SubjectNameRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "UnitNameRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "UnitResult": {
        "required": [
          "id",
          "subjectId",
          "name",
          "order",
          "lessonCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subjectId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "order": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "lessonCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "UpdateLessonRequest": {
        "required": [
          "name",
          "explanation",
          "summary",
          "videoUrl",
          "objectives"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "explanation": {
            "type": [
              "null",
              "string"
            ]
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "videoUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "objectives": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LessonObjectiveRequest"
            }
          }
        }
      },
...
    }
  },
...
}